+++ packages/create/src/generators/linting/templates/static/.eslintrc.js
@@ -1,0 +1,3 @@
+module.exports = {
+  extends: ['@open-wc/eslint-config', 'eslint-config-prettier'],
+};