+++ packages/demoing-storybook/demo/.storybook/config.js
@@ -1,4 +1,5 @@
-import { configure } from '@storybook/polymer';
+import { configure, addDecorator } from '@storybook/polymer';
+import { withA11y } from '@storybook/addon-a11y';
 import '@storybook/addon-console';
 
 const req = require.context('../stories', true, /\.stories\.js$/);
@@ -6,4 +7,5 @@ function loadStories() {
   req.keys().forEach(filename => req(filename));
 }
 
+addDecorator(withA11y);
 configure(loadStories, module);