+++ packages/es-dev-server/demo/index-rewrite/es-dev-server.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+  port: 8000,
+  customMiddlewares: [
+    function rewriteIndex(ctx, next) {
+      if (ctx.url === '/' || ctx.url === '/index.html') {
+        ctx.url = '/demo/index-rewrite/index.html';
+      }
+
+      return next();
+    },
+  ],
+};