+++ packages/building-utils/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.5.0](https://github.com/open-wc/open-wc/compare/@open-wc/building-utils@2.4.3...@open-wc/building-utils@2.5.0) (2019-07-22)
+
+
+### Features
+
+* **es-dev-server:** handle inline modules ([#596](https://github.com/open-wc/open-wc/issues/596)) ([49e4013](https://github.com/open-wc/open-wc/commit/49e4013))
+
+
+
+
+
+## [2.4.3](https://github.com/open-wc/open-wc/compare/@open-wc/building-utils@2.4.2...@open-wc/building-utils@2.4.3) (2019-07-19)
+
+**Note:** Version bump only for package @open-wc/building-utils
+
+
+
+
+
 ## [2.4.2](https://github.com/open-wc/open-wc/compare/@open-wc/building-utils@2.4.1...@open-wc/building-utils@2.4.2) (2019-07-15)