+++ packages/demoing-storybook/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0](https://github.com/open-wc/open-wc/compare/@open-wc/demoing-storybook@0.2.5...@open-wc/demoing-storybook@0.3.0) (2019-07-22)
+
+
+### Features
+
+* **demoing-storybook:** add a11y addon to package and generator ([#605](https://github.com/open-wc/open-wc/issues/605)) ([60bd5f7](https://github.com/open-wc/open-wc/commit/60bd5f7))
+
+
+
+
+
 ## [0.2.5](https://github.com/open-wc/open-wc/compare/@open-wc/demoing-storybook@0.2.4...@open-wc/demoing-storybook@0.2.5) (2019-06-17)