+++ packages/testing-helpers/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.18](https://github.com/open-wc/open-wc/compare/@open-wc/testing-helpers@1.0.17...@open-wc/testing-helpers@1.0.18) (2019-07-22)
+
+**Note:** Version bump only for package @open-wc/testing-helpers
+
+
+
+
+
+## [1.0.17](https://github.com/open-wc/open-wc/compare/@open-wc/testing-helpers@1.0.16...@open-wc/testing-helpers@1.0.17) (2019-07-22)
+
+**Note:** Version bump only for package @open-wc/testing-helpers
+
+
+
+
+
+## [1.0.16](https://github.com/open-wc/open-wc/compare/@open-wc/testing-helpers@1.0.15...@open-wc/testing-helpers@1.0.16) (2019-07-19)
+
+**Note:** Version bump only for package @open-wc/testing-helpers
+
+
+
+
+
+## [1.0.15](https://github.com/open-wc/open-wc/compare/@open-wc/testing-helpers@1.0.14...@open-wc/testing-helpers@1.0.15) (2019-07-17)
+
+**Note:** Version bump only for package @open-wc/testing-helpers
+
+
+
+
+
 ## [1.0.14](https://github.com/open-wc/open-wc/compare/@open-wc/testing-helpers@1.0.13...@open-wc/testing-helpers@1.0.14) (2019-07-17)
 
 **Note:** Version bump only for package @open-wc/testing-helpers