+++ packages/testing/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.6](https://github.com/open-wc/open-wc/compare/@open-wc/testing@2.0.5...@open-wc/testing@2.0.6) (2019-07-22)
+
+**Note:** Version bump only for package @open-wc/testing
+
+
+
+
+
+## [2.0.5](https://github.com/open-wc/open-wc/compare/@open-wc/testing@2.0.4...@open-wc/testing@2.0.5) (2019-07-22)
+
+**Note:** Version bump only for package @open-wc/testing
+
+
+
+
+
+## [2.0.4](https://github.com/open-wc/open-wc/compare/@open-wc/testing@2.0.3...@open-wc/testing@2.0.4) (2019-07-19)
+
+**Note:** Version bump only for package @open-wc/testing
+
+
+
+
+
+## [2.0.3](https://github.com/open-wc/open-wc/compare/@open-wc/testing@2.0.2...@open-wc/testing@2.0.3) (2019-07-17)
+
+**Note:** Version bump only for package @open-wc/testing
+
+
+
+
+
 ## [2.0.2](https://github.com/open-wc/open-wc/compare/@open-wc/testing@2.0.1...@open-wc/testing@2.0.2) (2019-07-17)
 
 **Note:** Version bump only for package @open-wc/testing