+++ packages/webpack-index-html-plugin/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.9](https://github.com/open-wc/open-wc/compare/@open-wc/webpack-index-html-plugin@1.2.8...@open-wc/webpack-index-html-plugin@1.2.9) (2019-07-22)
+
+**Note:** Version bump only for package @open-wc/webpack-index-html-plugin
+
+
+
+
+
+## [1.2.8](https://github.com/open-wc/open-wc/compare/@open-wc/webpack-index-html-plugin@1.2.7...@open-wc/webpack-index-html-plugin@1.2.8) (2019-07-19)
+
+**Note:** Version bump only for package @open-wc/webpack-index-html-plugin
+
+
+
+
+
 ## [1.2.7](https://github.com/open-wc/open-wc/compare/@open-wc/webpack-index-html-plugin@1.2.6...@open-wc/webpack-index-html-plugin@1.2.7) (2019-07-15)